requirements.txt
Flask==2.0.3
requests>=2.28.0
watchdog==2.1.6
tmdbv3api>=1.9.0,<2.0.0
pytest==6.2.5
python-dotenv>=0.20.0